Cleaned up parentSidebar and moved code over to the parentHome index.js
